Fix a bug when utilizing additional search paths for tags files.
Issue manifested as a python exception trying to open a directory as a file when 'extra_tag_paths' is non-empty. This was caused by the code that examines the search paths to assume all searchs paths lead to a fully qualified file.

The fix is to do exaclty that by adding the tag_file setting to the end.  This will need to be changed in order to support multiple tag filenames.