Changes to be committed: 	modified:   ctags.py
Added checks to prevent the search for tags from returning the blank
line at the end of the tag file as a tag.

This fixes a bug that is manifested when searching for tags in the file
(Alt+S) that happens to appear last in the '.tags_sorted_by_file' file.

Solves issue #118