Fix CTags always searches the tag file in the first top folder in a project
For example, I have two top folders in my project.

In sidebar, it looks like

   + dir_A
   + dir_B

Ctags always only search the tag file (.tags) in dir_A
even if you are editing a file in dir_B.

This commit fixes that. CTags now searches the folder
which contains the file we are editing.